REF/TST: collect Index join tests (#32171)
