+++ pandas/tests/indexes/timedeltas/test_timedelta.py
@@ -91,27 +91,6 @@ class TestTimedeltaIndex(DatetimeLike):
         tm.assert_numpy_array_equal(arr, exp_arr)
         tm.assert_index_equal(idx, idx3)
 
-    def test_join_self(self, join_type):
-        index = timedelta_range("1 day", periods=10)
-        joined = index.join(index, how=join_type)
-        tm.assert_index_equal(index, joined)
-
-    def test_does_not_convert_mixed_integer(self):
-        df = tm.makeCustomDataframe(
-            10,
-            10,
-            data_gen_f=lambda *args, **kwargs: randn(),
-            r_idx_type="i",
-            c_idx_type="td",
-        )
-        str(df)
-
-        cols = df.columns.join(df.index, how="outer")
-        joined = cols.join(df.columns)
-        assert cols.dtype == np.dtype("O")
-        assert cols.dtype == joined.dtype
-        tm.assert_index_equal(cols, joined)
-
     def test_sort_values(self):
 
         idx = TimedeltaIndex(["4d", "1d", "2d"])
@@ -180,16 +159,6 @@ class TestTimedeltaIndex(DatetimeLike):
             TypeError, match=(f"unhashable type: {repr(type(index).__name__)}")
         ):
             hash(index)
-
-    def test_append_join_nondatetimeindex(self):
-        rng = timedelta_range("1 days", periods=10)
-        idx = Index(["a", "b", "c", "d"])
-
-        result = rng.append(idx)
-        assert isinstance(result[0], Timedelta)
-
-        # it works
-        rng.join(idx, how="outer")
 
     def test_append_numpy_bug_1681(self):
